misc: promote walkable heights to int32_t

+++ src/trx/game/objects/general/bridge.c
@@ -104,8 +104,8 @@ static int32_t M_GetOffsetHeight(const ITEM *const item, const XYZ_32 pos)
     }
 }
 
-static int16_t M_GetFloorHeight(
-    const ITEM *const item, const XYZ_32 pos, const int16_t height)
+static int32_t M_GetFloorHeight(
+    const ITEM *const item, const XYZ_32 pos, const int32_t height)
 {
     if (g_Config.gameplay.fix_bridge_collision && !M_IsSameSector(pos, item)) {
         return height;
@@ -123,8 +123,8 @@ static int16_t M_GetFloorHeight(
     return offset_height;
 }
 
-static int16_t M_GetCeilingHeight(
-    const ITEM *const item, const XYZ_32 pos, const int16_t height)
+static int32_t M_GetCeilingHeight(
+    const ITEM *const item, const XYZ_32 pos, const int32_t height)
 {
     if (g_Config.gameplay.fix_bridge_collision && !M_IsSameSector(pos, item)) {
         return height;
